requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-telegram-bot[job-queue]>=21.0
anthropic>=0.28.0
elevenlabs>=1.2.0
httpx>=0.27.0
requests>=2.32.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
